requirements.txt
amqp==2.3.2
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.4
celery==4.2.1
colorama==0.4.0
Django==2.1.2
django-celery-results==1.0.1
django-db-signals==0.1.1
django-filter==2.0.0
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==0.9.2
kombu==4.2.1
Markdown==3.0.1
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pytest==3.9.3
python-dateutil==2.7.5
pytz==2018.5
six==1.11.0
text-unidecode==1.2
vine==1.1.4
